+++ modules/kibana/src/internalClusterTest/java/org/elasticsearch/kibana/KibanaThreadPoolIT.java
@@ -12,8 +12,9 @@ import org.elasticsearch.action.bulk.BulkResponse;
 import org.elasticsearch.action.support.WriteRequest;
 import org.elasticsearch.client.internal.Client;
 import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.indices.SystemIndexThreadPoolTests;
+import org.elasticsearch.indices.SystemIndexThreadPoolTestCase;
 import org.elasticsearch.plugins.Plugin;
+import org.elasticsearch.test.ESIntegTestCase;
 
 import java.util.Collection;
 import java.util.Map;
@@ -22,7 +23,7 @@ import java.util.Set;
 import static org.elasticsearch.test.hamcrest.ElasticsearchAssertions.assertHitCount;
 import static org.elasticsearch.test.hamcrest.ElasticsearchAssertions.assertNoFailures;
 
-public class KibanaThreadPoolTests extends SystemIndexThreadPoolTests {
+public class KibanaThreadPoolIT extends SystemIndexThreadPoolTestCase {
 
     @Override
     protected Collection<Class<? extends Plugin>> nodePlugins() {
@@ -32,11 +33,11 @@ public class KibanaThreadPoolTests extends SystemIndexThreadPoolTests {
     public void testKibanaThreadPool() {
         runWithBlockedThreadPools(() -> {
             // index documents
-            String idToDelete = client().prepareIndex(".kibana").setSource(Map.of("foo", "delete me!")).get().getId();
-            String idToUpdate = client().prepareIndex(".kibana").setSource(Map.of("foo", "update me!")).get().getId();
+            String idToDelete = ESIntegTestCase.client().prepareIndex(".kibana").setSource(Map.of("foo", "delete me!")).get().getId();
+            String idToUpdate = ESIntegTestCase.client().prepareIndex(".kibana").setSource(Map.of("foo", "update me!")).get().getId();
 
             // bulk index, delete, and update
-            Client bulkClient = client();
+            Client bulkClient = ESIntegTestCase.client();
             BulkResponse response = bulkClient.prepareBulk(".kibana")
                 .add(bulkClient.prepareIndex(".kibana").setSource(Map.of("foo", "search me!")))
                 .add(bulkClient.prepareDelete(".kibana", idToDelete))
@@ -46,7 +47,7 @@ public class KibanaThreadPoolTests extends SystemIndexThreadPoolTests {
             assertNoFailures(response);
 
             // match-all search
-            assertHitCount(client().prepareSearch(".kibana").setQuery(QueryBuilders.matchAllQuery()), 2);
+            assertHitCount(ESIntegTestCase.client().prepareSearch(".kibana").setQuery(QueryBuilders.matchAllQuery()), 2);
         });
     }
 }
