+++ test/framework/src/main/java/org/elasticsearch/indices/SystemIndexThreadPoolTestCase.java
